requirements.txt
google-genai
groq
openai
python-dotenv
httpx
html2text
rich
imap_tools
dnspython
requests
